Readfile: Avoid broad exception catch
The previous code was catching all the exceptions, hiding the important
data about the cause of the failure.

Now it only catches the known common error cause: file not found. The
error message was updated accordingly.

All other kinds of errors will raise a failure and show the normal
traceback, so the user can have a clue on what went wrong.