Readfile: Avoid too broad exception catch
As currently done in Savefile(), only OS errors are the "safe" errors to
catch when trying to read a file. All other kinds of errors should raise
a failure and show the normal traceback, because they are unexpected.

For consistency, also changed Savefile() to use OSError. Both should
be equivalent: https://stackoverflow.com/q/29347790/