Force UTF-8 enconding when reading/writing files (#221)
* Force UTF-8 enconding when reading/writing files

Since e4e56d9 txt2tags dropped `--encoding` and is only supporting UTF-8
as the input and output file encoding. In this commit this is now
explicitly enforced in the reading/writing files routines.

Sometimes, depending on the OS and the environment settings, UTF-8 is
not the default encoding for open(). For example, in Windows it could be
cp1252. In that case, a valid UTF-8 input file is read as cp1252 when no
encoding is specified in open(). Learn more at
https://stackoverflow.com/a/42070962/.

Note that for Python 2, now all the file contents is available as a
Unicode string (u"foo" instead of "foo"). So explicit checks for
`isinstance(str)` will fail, as well as code that assumes `str`.

Fixes #212

* Add new test mode: _grep()

Using this new mode, the OK template now can be a snippet of the test
output, allowing for partial matches.

It is handy to avoid smaller differences between the output from
different Python versions.