[ attempt at #228 ] definition lists in markdown
I tried to change the generated markdown for definition lists to match
PHP Markdown Extra / PanDoc.

Currently, the -t md just reproduces the t2t format of definition
lists, which is not matching any flavor of markdown, AFAICT.

What I am trying to produce is in the patch to sample.md.
(This does not handle nested lists correctly, but so doesn't the
status quo either.)

However, I do not get there because the framework inserts a newline
between the `deflistItem2Open` tag and the description text.

To fix this, the framework has to be changed, it is not sufficient to
tinker with the values of deflistItem*, it seems.

Run test/run.py to see the difference between what my patch achieves
and where we want to get to.