issue #227 md table: correct title cell separators
class TableMaster function dump has a hack for tex output as well, so I saw not much harm to add a hack for md table titles too. An md table row is marked as title row, by adding a row beneath it with as much '---' cells as there are cells in the title row.